requirements.txt
alembic
click
click-completion
flask
flask-classful
flask-httpauth
flask-socketio
werkzeug==2.1.2 # fixes a problem
opencage
pluggy
pbr
pyyaml
# Allowing a newer version can lead to a conflict with
# requests.
pytz
requests
thesmuggler
update_checker
sqlalchemy
psycopg2>=2.9.0
GeoAlchemy2
oslo.config
oslo.log
yaspin
tabulate
pygments
sentry-sdk
blinker
mapbox
geojson
pylibmc
rich
# For sqlalchemy
dogpile.cache
ratelimit
paho-mqtt>=2.0.0
# Generic caching
cachetools
geopy
loguru
